+++ packages/to-json-schema/src/converters/convertSchema/convertSchema.test.ts
@@ -882,7 +882,7 @@ describe('convertSchema', () => {
           createContext()
         )
       ).toStrictEqual({
-        anyOf: [
+        oneOf: [
           {
             type: 'object',
             properties: {